admin party completed

+++ src/main/java/com/gachaapi/Repository/PlayerRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PlayerRepository extends JpaRepository<Player, Long> {
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
     Optional<Player> findByNick(String nick);
     boolean existsByNick(String nick);
 }
